+++ src/main/java/io/mycat/buffer/DirectByteBufferPool.java
@@ -6,9 +6,6 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.atomic.AtomicInteger;
 
-import io.mycat.MycatServer;
-import io.mycat.buffer.handler.RecycleThread;
-import io.mycat.util.NameableExecutor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -30,17 +27,13 @@ public class DirectByteBufferPool implements BufferPool{
     private final  int pageSize;
     private final short pageCount;
     private final int conReadBuferChunk ;
-    public static BlockingQueue<ByteBuffer> noBlockingQueue = new LinkedBlockingQueue<ByteBuffer>();
-    //回收个数
-	static AtomicInteger count = new AtomicInteger(0);
-	//申请个数
-	public static AtomicInteger apply = new AtomicInteger(0);
-    /**
+      
+     /**
      * 记录对线程ID->该线程的所使用Direct Buffer的size
      */
     private final ConcurrentHashMap<Long,Long> memoryUsage;
 
-    public DirectByteBufferPool(int pageSize, short chunkSize, short pageCount, int conReadBuferChunk,final NameableExecutor businessExecutor) {
+    public DirectByteBufferPool(int pageSize, short chunkSize, short pageCount,int conReadBuferChunk) {
         allPages = new ByteBufferPage[pageCount];
         this.chunkSize = chunkSize;
         this.pageSize = pageSize;
@@ -51,11 +44,6 @@ public class DirectByteBufferPool implements BufferPool{
             allPages[i] = new ByteBufferPage(ByteBuffer.allocateDirect(pageSize), chunkSize);
         }
         memoryUsage = new ConcurrentHashMap<>();
-
-        //启动回收直接内存线程
-        if(businessExecutor !=null)
-             businessExecutor.
-                     execute(new RecycleThread("RecycleBufferThread", allPages,memoryUsage));
     }
 
     public BufferArray allocateArray() {
@@ -105,13 +93,33 @@ public class DirectByteBufferPool implements BufferPool{
     }
 
     public void recycle(ByteBuffer theBuf) {
-     	LOGGER.debug("recyclequeue process number :"+count.incrementAndGet());
-     	if(theBuf !=null && (!(theBuf instanceof DirectBuffer) )){
+      	if(theBuf !=null && (!(theBuf instanceof DirectBuffer) )){
     		theBuf.clear();
     		return;
          }
-    	//huangyiming add  异步队列回收堆外内存
-    	noBlockingQueue.offer(theBuf);
+ 		 
+		final long size = theBuf.capacity();
+
+		boolean recycled = false;
+		DirectBuffer thisNavBuf = (DirectBuffer) theBuf;
+		int chunkCount = theBuf.capacity() / chunkSize;
+		DirectBuffer parentBuf = (DirectBuffer) thisNavBuf.attachment();
+		int startChunk = (int) ((thisNavBuf.address() - parentBuf.address()) / chunkSize);
+		for (int i = 0; i < allPages.length; i++) {
+			if ((recycled = allPages[i].recycleBuffer((ByteBuffer) parentBuf, startChunk,
+					chunkCount) == true)) {
+				break;
+			}
+		}
+		final long threadId = Thread.currentThread().getId();
+
+		if (memoryUsage.containsKey(threadId)) {
+			memoryUsage.put(threadId, memoryUsage.get(threadId) - size);
+		}
+		if (recycled == false) {
+			LOGGER.warn("warning ,not recycled buffer " + theBuf);
+		}
+	
     }
 
     private ByteBuffer allocateBuffer(int theChunkCount, int startPage, int endPage) {
